CAT-2359 Refactor UserListBricks to use the same Touch / Item Selectedlisteners, fix spinner selection for UserListBricks and UserVariableBricks, fix corresponding tests
